+++ src/main/java/de/muenchen/relationalimpl/ZaehlstelleIndexImpl.java
@@ -1,19 +1,17 @@
 package de.muenchen.relationalimpl;
 
-import de.muenchen.dave.configuration.CachingConfiguration;
 import de.muenchen.dave.domain.elasticsearch.Zaehlstelle;
 import de.muenchen.dave.domain.elasticsearch.Zaehlung;
-import de.muenchen.dave.domain.mapper.FahrbeziehungMapper;
-import de.muenchen.dave.domain.mapper.ZaehlstelleMapper;
-import de.muenchen.dave.domain.mapper.ZaehlungMapper;
 import de.muenchen.dave.repositories.elasticsearch.ZaehlstelleIndex;
 import de.muenchen.dave.repositories.relationaldb.ZaehlstelleRepository;
 import de.muenchen.dave.repositories.relationaldb.ZaehlungRepository;
+import de.muenchen.relationalimpl.mapper.FahrbeziehungRelationalMapper;
+import de.muenchen.relationalimpl.mapper.ZaehlstelleRelationalMapper;
+import de.muenchen.relationalimpl.mapper.ZaehlungRelationalMapper;
 import java.util.List;
 import java.util.Optional;
 import java.util.UUID;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.cache.annotation.CacheEvict;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
@@ -26,17 +24,17 @@ public class ZaehlstelleIndexImpl implements ZaehlstelleIndex {
 
     private final ZaehlungRepository zaehlungRepository;
 
-    private final ZaehlstelleMapper zaehlstelleMapper;
+    private final ZaehlstelleRelationalMapper zaehlstelleMapper;
 
-    private final ZaehlungMapper zaehlungMapper;
+    private final ZaehlungRelationalMapper zaehlungMapper;
 
-    private final FahrbeziehungMapper fahrbeziehungMapper;
+    private final FahrbeziehungRelationalMapper fahrbeziehungMapper;
 
     public ZaehlstelleIndexImpl(final ZaehlstelleRepository zaehlstelleRepository,
             final ZaehlungRepository zaehlungRepository,
-            final ZaehlstelleMapper zaehlstelleMapper,
-            final ZaehlungMapper zaehlungMapper,
-            final FahrbeziehungMapper fahrbeziehungMapper) {
+            final ZaehlstelleRelationalMapper zaehlstelleMapper,
+            final ZaehlungRelationalMapper zaehlungMapper,
+            final FahrbeziehungRelationalMapper fahrbeziehungMapper) {
         this.zaehlstelleRepository = zaehlstelleRepository;
         this.zaehlungRepository = zaehlungRepository;
         this.zaehlstelleMapper = zaehlstelleMapper;
@@ -44,55 +42,25 @@ public class ZaehlstelleIndexImpl implements ZaehlstelleIndex {
         this.fahrbeziehungMapper = fahrbeziehungMapper;
     }
 
-    @CacheEvict(
-            value = { CachingConfiguration.SUCHE_ERHEBUNGSSTELLE, CachingConfiguration.SUCHE_ERHEBUNGSSTELLE_DATENPORTAL,
-                    CachingConfiguration.LADE_BELASTUNGSPLAN_DTO, CachingConfiguration.LADE_PROCESSED_ZAEHLDATEN,
-                    CachingConfiguration.LADE_ZAEHLDATEN_ZEITREIHE_DTO, CachingConfiguration.READ_ZAEHLSTELLE_DTO },
-            allEntries = true
-    )
     public void deleteAll() {
         zaehlstelleRepository.deleteAll();
     }
 
-    @CacheEvict(
-            value = { CachingConfiguration.SUCHE_ERHEBUNGSSTELLE, CachingConfiguration.SUCHE_ERHEBUNGSSTELLE_DATENPORTAL,
-                    CachingConfiguration.LADE_BELASTUNGSPLAN_DTO, CachingConfiguration.LADE_PROCESSED_ZAEHLDATEN,
-                    CachingConfiguration.LADE_ZAEHLDATEN_ZEITREIHE_DTO, CachingConfiguration.READ_ZAEHLSTELLE_DTO },
-            allEntries = true
-    )
     public void deleteAll(Iterable<? extends Zaehlstelle> var1) {
         Iterable<de.muenchen.dave.domain.analytics.Zaehlstelle> analyticsList = zaehlstelleMapper.elasticlist2analyticslist(var1);
         zaehlstelleRepository.deleteAll(analyticsList);
     }
 
-    @CacheEvict(
-            value = { CachingConfiguration.SUCHE_ERHEBUNGSSTELLE, CachingConfiguration.SUCHE_ERHEBUNGSSTELLE_DATENPORTAL,
-                    CachingConfiguration.LADE_BELASTUNGSPLAN_DTO, CachingConfiguration.LADE_PROCESSED_ZAEHLDATEN,
-                    CachingConfiguration.LADE_ZAEHLDATEN_ZEITREIHE_DTO, CachingConfiguration.READ_ZAEHLSTELLE_DTO },
-            allEntries = true
-    )
     public void deleteById(String var1) {
         zaehlstelleRepository.deleteById(UUID.fromString(var1));
     }
 
-    @CacheEvict(
-            value = { CachingConfiguration.SUCHE_ERHEBUNGSSTELLE, CachingConfiguration.SUCHE_ERHEBUNGSSTELLE_DATENPORTAL,
-                    CachingConfiguration.LADE_BELASTUNGSPLAN_DTO, CachingConfiguration.LADE_PROCESSED_ZAEHLDATEN,
-                    CachingConfiguration.LADE_ZAEHLDATEN_ZEITREIHE_DTO, CachingConfiguration.READ_ZAEHLSTELLE_DTO },
-            allEntries = true
-    )
     public void delete(Zaehlstelle var1) {
         de.muenchen.dave.domain.analytics.Zaehlstelle zs = zaehlstelleMapper.elastic2analytics(new de.muenchen.dave.domain.analytics.Zaehlstelle(), var1,
                 zaehlungMapper, fahrbeziehungMapper);
         zaehlstelleRepository.delete(zs);
     }
 
-    @CacheEvict(
-            value = { CachingConfiguration.SUCHE_ERHEBUNGSSTELLE, CachingConfiguration.SUCHE_ERHEBUNGSSTELLE_DATENPORTAL,
-                    CachingConfiguration.LADE_BELASTUNGSPLAN_DTO, CachingConfiguration.LADE_PROCESSED_ZAEHLDATEN,
-                    CachingConfiguration.LADE_ZAEHLDATEN_ZEITREIHE_DTO, CachingConfiguration.READ_ZAEHLSTELLE_DTO },
-            allEntries = true
-    )
     public Zaehlstelle save(Zaehlstelle var1) {
         if (var1 == null) {
             return null;
@@ -111,6 +79,17 @@ public class ZaehlstelleIndexImpl implements ZaehlstelleIndex {
         //zaehlstelleEntity.setZaehlungen((List<de.muenchen.dave.domain.analytics.Zaehlung>) zaehlungen);
         zaehlstelleEntity = zaehlstelleRepository.save(zaehlstelleEntity);
         return zaehlstelleMapper.analytics2elastic(zaehlstelleEntity);
+    }
+
+    public Iterable<Zaehlstelle> saveAll(Iterable<Zaehlstelle> var1) {
+        if (var1 == null) {
+            return null;
+        }
+        List<Zaehlstelle> zaehlstellenList = new java.util.ArrayList<>();
+        for (Zaehlstelle zaehlstelle : var1) {
+            zaehlstellenList.add(this.save(zaehlstelle));
+        }
+        return zaehlstellenList;
     }
 
     public Optional<Zaehlstelle> findById(String var1) {
